P language extension: The character size can be expressed as a persentage of   the character size of another element, typically the parent element, e.g:   Size: Enclosing * 80 %;

+++ amaya/HTMLform.c
@@ -114,7 +114,7 @@ STRING              orig;
    int                 lg;
 
    lg = ustrlen (orig) + 1;
-   if (ustrlen (buffer) + lg > lgbuffer)
+   if ((int)ustrlen (buffer) + lg > lgbuffer)
      {
 	/* it is necessary to extend the buffer */
 	if (lg < PARAM_INCREMENT)
